Refactor: CategoryLearningProgress Flyway 버전 수정

+++ user-api/src/main/resources/db/migration/V1.18.0__create.sql
@@ -2,8 +2,8 @@ CREATE TABLE IF NOT EXISTS `user_category_progress`
 (
     `user_id`               BIGINT NOT NULL,
     `category_id`           BIGINT NOT NULL,
-    `total_learn_count`     INT    NOT NULL DEFAULT 0,
-    `completed_learn_count` INT    NOT NULL DEFAULT 0,
+    `total_learn_count`     BIGINT NOT NULL DEFAULT 0,
+    `completed_learn_count` BIGINT NOT NULL DEFAULT 0,
     PRIMARY KEY (`user_id`, `category_id`)
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8mb4
